backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
